chore: move cli package to internal
