+++ pkg/cmd/list/list.go
@@ -12,8 +12,8 @@ import (
 	"fmt"
 
 	"github.com/spf13/cobra"
+	"github.com/vchain-us/vcn/internal/cli"
 	"github.com/vchain-us/vcn/pkg/api"
-	"github.com/vchain-us/vcn/pkg/cli"
 	"github.com/vchain-us/vcn/pkg/store"
 )
